refactored all actions to be in an actions file for their respective scope to break up the giant file that was Request
