+++ client/src/context/actions/discussionActions.js
@@ -1,0 +1,120 @@
+import { AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, RNS3, ImagePicker, axios, local, postgres } from './constants';
+
+// handles aws image uploading
+export const uploadImage = dispatch => {
+    ImagePicker.showImagePicker({}, response => {
+        /*response returns an object with all of the information about the selected image.
+        returns data, fileName, fileSize, height, isVertical, latitude, longitude, origURL,
+        timestamp, type, uri, width.
+        */
+        // extract this data for the file upload
+        const file = {
+            uri: response.uri,
+            name: response.fileName,
+            type: 'image/png'
+        };
+        // s3 configurations
+        const config = {
+            keyPrefix: 's3/',
+            bucket: 'lambdasocialbucket',
+            region: 'us-east-1',
+            accessKey: AWS_ACCESS_KEY_ID,
+            secretKey: AWS_SECRET_ACCESS_KEY,
+            successActionStatus: 201
+        };
+        dispatch({ type: 'SENDING_IMAGE' });
+        RNS3.put(file, config)
+            .then(response => {
+                if (response.status === 403) {
+                    return dispatch({
+                        type: 'IMAGE_FAILED',
+                        payload: 'Failed to upload image to S3'
+                    });
+                } else {
+                    /*
+                    response will come back looking like this, we'll want
+                    the location for the POST request to make a discussion.
+                        location: "https://lambdasocialbucket.s3.amazonaws.com/s3%2FIMG_0111.HEIC"
+                  */
+                    return dispatch({
+                        type: 'IMAGE_SUCCESS',
+                        payload: response.body.postResponse.location
+                    });
+                }
+            })
+            .catch(err => {
+                console.log(err);
+                return dispatch({ type: 'IMAGE_FAILED', payload: err });
+            });
+    });
+};
+
+export const getDiscussions = async (query, dispatch) => {
+    // handle loading state
+    const q = new URLSearchParams({ sort: query });
+    dispatch({ type: 'TOP_DISCUSSIONS_FETCHING', payload: true });
+    try {
+        const res = await axios.get(`${postgres}/discussions/?${q.toString()}`);
+        return dispatch({ type: 'TOP_DISCUSSIONS_FETCHED', payload: res.data });
+    } catch (err) {
+        console.log(err);
+        return dispatch({ type: 'TOP_DISCUSSIONS_FAILED', payload: err });
+    }
+};
+
+export const getDiscussionsForSub = async (id, dispatch) => {
+    try {
+        await dispatch({ type: 'DISCUSSIONS_FETCHING', payload: true });
+        const res = await axios.get(`${postgres}/discussions/s/${id}`);
+        return dispatch({ type: 'DISCUSSIONS_FETCHED', payload: res.data });
+    } catch (err) {
+        console.log(err);
+        return dispatch({ type: 'DISCUSSIONS_FAILED', payload: err });
+    }
+};
+
+export const getRecentDiscussions = async (id, dispatch) => {
+    const body = {
+        id
+    }
+    try {
+        await dispatch({ type: 'DISCUSSIONS_FETCHING', payload: true });
+        const res = await axios.post(`${postgres}/discussions/recent`, body);
+        return dispatch({
+            type: 'DISCUSSIONS_FETCHED',
+            payload: res.data
+        });
+    } catch (err) {
+        console.log(err);
+        return dispatch({ type: 'DISCUSSIONS_FAILED', payload: err });
+    }
+};
+
+export const addDiscussion = async (body, dispatch, nav) => {
+    dispatch({ type: 'DISCUSSIONS_FETCHING' })
+    const apiBody = {
+        title: body.title,
+        content: body.content,
+        image: body.image,
+        creater_id: body.creater_id,
+        subtopic_id: body.subtopic_id,
+        hashtags: body.hashtag,
+    };
+
+    try {
+        let res = await axios.post(`${postgres}/discussions/create`, apiBody);
+
+        let followup = await dispatch({
+            type: 'CREATED_DISCUSSION',
+            payload: res.data
+        });
+
+        return { res, followup };
+    } catch (err) {
+        dispatch({
+            type: 'CREATE_DISCUSSION_FAILED',
+            payload: res.response.data
+        })
+        console.log(err);
+    }
+};
